chore(modules): Organize utils
