+++ src/auth/auth.ts
@@ -3,8 +3,7 @@ import { Observable } from 'rxjs/Observable';
 import { Observer } from 'rxjs/Observer';
 import { ReplaySubject } from 'rxjs/ReplaySubject';
 import { FirebaseApp, FirebaseAuthConfig, WindowLocation } from '../tokens';
-import { isPresent } from '../utils/utils';
-import * as utils from '../utils/utils';
+import * as utils from '../utils';
 import {
   authDataToAuthState,
   AuthBackend,
@@ -84,12 +83,12 @@ export class AngularFireAuth extends ReplaySubject<FirebaseAuthState> {
     }
     config = this._mergeConfigs(config);
 
-    if (!isPresent(config.method)) {
+    if (!utils.isPresent(config.method)) {
       return this._reject('You must provide a login method');
     }
     let providerMethods = [AuthMethods.Popup, AuthMethods.Redirect, AuthMethods.OAuthToken];
     if (providerMethods.indexOf(config.method) != -1) {
-      if (!isPresent(config.provider)) {
+      if (!utils.isPresent(config.provider)) {
         return this._reject('You must include a provider to use this auth method.');
       }
     }
